+++ common/addresses/AddrKey.sol
@@ -19,4 +19,6 @@ library AddrKey {
     string constant public wS = "wS";
     string constant public stS = "stS";
     string constant public ODOS_ROUTER = "ODOS_ROUTER";
+    string constant public DAO = "DAO";
+    string constant public scUSD = "scUSD";
 }